change message structure to a chan

+++ exchange/binance/order.go
@@ -12,7 +12,7 @@ import (
 )
 
 var (
-	TradeDescribers = map[int]*TradeDescriber{}
+	TradeDescribers = map[int64]*TradeDescriber{}
 )
 
 type PreparedOrder struct {
